backend/requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
slack_sdk>=3.21.0
streamlit>=1.28.0
pandas>=1.5.0
supabase>=0.7.1